+++ netzob/src/netzob/Fuzzing/DeterministIntegerMutator.py
@@ -65,7 +65,7 @@ class DeterministIntegerMutator(DomainMutator):
         between the min and max values of the domain.
         If set to :attr:`MutatorInterval.FULL_INTERVAL <netzob.Fuzzing.DomainMutator.MutatorInterval.FULL_INTERVAL>`, the values will be generate in
         [0, 2^N-1], where N is the bitsize (storage) of the field.
-        If it is an tuple of integers (min, max), the values will be generate
+        If it is a tuple of integers (min, max), the values will be generate
         between min and max.
         Default value is :attr:`MutatorInterval.DEFAULT_INTERVAL <netzob.Fuzzing.DomainMutator.MutatorInterval.DEFAULT_INTERVAL>`.
     :param mode: If set to :attr:`MutatorMode.GENERATE <netzob.Fuzzing.DomainMutator.MutatorMode.GENERATE>`, :meth:`generate` will be