+++ netzob/test/src/test_netzob/test_Common/test_Type/test_Endianness.py
@@ -25,14 +25,14 @@
 #|             Supélec, http://www.rennes.supelec.fr/ren/rd/cidre/           |
 #+---------------------------------------------------------------------------+
 
-from netzob.Common.Type.Endianess import Endianess
+from netzob.Common.Type.Endianness import Endianness
 from common.NetzobTestCase import NetzobTestCase
 
 
-class test_Endianess(NetzobTestCase):
+class test_Endianness(NetzobTestCase):
 
     def test_BIG(self):
-        self.assertEqual(Endianess.BIG, "big-endian")
+        self.assertEqual(Endianness.BIG, "big-endian")
 
     def test_LITTLE(self):
-        self.assertEqual(Endianess.LITTLE, "little-endian")
+        self.assertEqual(Endianness.LITTLE, "little-endian")